Fix the leaking callbacks list in SwiDebugView by implementing our own erase() that also clears the callbacks.
Just adding an erase() on SwiDebugView did not do it, as the __getattr__
redirected it to the erase() on the backing view. (Not clear why, as
it seems that it is explicitly is not supposed to be called unless
the method is not already on self.)

So I removed the pass-through mechanism in __getattr__ and implemented
explicit passthroughs for everything we use on View. That's more
encapsulated and explicit anyway.

There remains a bug where only a partial erase occurs. It should be
removing some of the callbacks:
self.view.erase(edit, sublime.Region(erase_to, self.view.size()))